moved app component into components directory, and tweaked README
