+++ src/app/app.components.ts
@@ -1,5 +1,5 @@
 import { HomeComponent } from './components/home/home.component';
-import { AppComponent } from './app.component';
+import { AppComponent } from './components/app/app.component';
 
 export const APP_COMPONENTS = [
     AppComponent,
