+++ src/app/app.routes.ts
@@ -2,6 +2,6 @@ import { Routes } from '@angular/router';
 
 import { HomeComponent } from './components/home/home.component';
 
-export const APP_Routes: Routes = [
+export const APP_ROUTES: Routes = [
     { path: '', component: HomeComponent }
 ];
